atom.xml: don't assume <updated> to be last updated post
The user may have triggered an manual regeneration, so it may not be
accurate to assume that the feed's last-modified date is the same as
that of the last updated post.

We change <updated> and <id>, but leave <rights> alone, since the
copyright year should rely on the post's update date.